refactor velocity processing responsibilities
